Fix RequestMetadata setup for incoming requests on IGO_NEW_REQUEST topic Added fail-safe check on MetadbRequest obj construction in unit tests
